+++ src/lambdaInit.h
@@ -1,32 +1,32 @@
 #define _USE_MATH_DEFINES
 
-int initZero(int Height,int Width, float **gFieldOut);
+int initZero(int Height,int Width, double **gFieldOut);
 
-int gradUtoLamb(int Height,int Width, float *gField,float **sFieldOut);
+int gradUtoLamb(int Height,int Width, double *gField,double **sFieldOut);
 
-int addSingleOseen(int nVortex,float *parVortex, float *x0, float *dx, 
-                   int Height,int Width, float **gFieldOut);
+int addSingleOseen(int nVortex,double *parVortex, double *x0, double *dx, 
+                   int Height,int Width, double **gFieldOut);
 
-int addUSingleOseen(int nVortex,float *parVortex, float *x0, float *dx, 
-                    int Height,int Width, float **uFieldOut);
+int addUSingleOseen(int nVortex,double *parVortex, double *x0, double *dx, 
+                    int Height,int Width, double **uFieldOut);
 
-int addConstXYShear(float *x0, float *dx,
+int addConstXYShear(double *x0, double *dx,
 	                int Height,int Width, 
-	                float v0y0,float **gFieldOut);
+	                double v0y0,double **gFieldOut);
 
-int initLambOseen2D(int nVortex,float *parVortex,
-                    float *x0, float *dx, int Height,int Width,
-                    float **sFieldOut);
+int initLambOseen2D(int nVortex,double *parVortex,
+                    double *x0, double *dx, int Height,int Width,
+                    double **sFieldOut);
 
-int initOseenShear2D(int nVortex,float *parVortex,
-                     float *x0, float *dx, int Height,int Width,
-                     float v0y0, float **sFieldOut);
+int initOseenShear2D(int nVortex,double *parVortex,
+                     double *x0, double *dx, int Height,int Width,
+                     double v0y0, double **sFieldOut);
 
-int addOseen2ndGrad(int nVortex,float *parVortex, float *x0, float *dx, 
-                    int Height,int Width, float **gFieldOut);
+int addOseen2ndGrad(int nVortex,double *parVortex, double *x0, double *dx, 
+                    int Height,int Width, double **gFieldOut);
 
-int s2ndGradUtoLamb(int nVortex,float *parVortex, float *x0, float *dx,
-                    int Height,int Width, float *gField,float *sField);
+int s2ndGradUtoLamb(int nVortex,double *parVortex, double *x0, double *dx,
+                    int Height,int Width, double *gField,double *sField);
 
-int sndGradUwFieldToLamb(int Height,int Width,float *gField,float *wField,
-                         float *sField);
+int sndGradUwFieldToLamb(int Height,int Width,double *gField,double *wField,
+                         double *sField);
